Split Straight PathLike implementation into own file
